Refactor reactor loop to improve connection teardown stability
Iterate through poll_fds backwards so swap-and-pop swaps in an element that has already been handled

Prioritize POLLERR, POLLHUP, and POLLNVAL checks to handle dead sockets before processing data

call close explicitly during client disconnection to prevent FD leaks

+++ src/server/tcp_server.hpp
@@ -75,9 +75,9 @@ private:
     // Reactor event loop
     void run_reactor();
     void handle_new_connection();
-    void handle_new_command(size_t& poll_fds_idx);
-    void handle_client_write(size_t& poll_fds_idx);
-    void handle_client_dc(size_t& poll_fds_idx);
+    void handle_new_command(int& poll_fds_idx);
+    void handle_client_write(int& poll_fds_idx);
+    void handle_client_dc(int& poll_fds_idx);
 
     // Thread pool
     size_t num_workers_{5};
